backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.0
python-multipart
minio>=7.0.0
PyPDF2>=3.0.0
PyExifTool>=0.5.0
httpx>=0.24.0
python-jose[cryptography]  # For JWT
passlib[bcrypt]           # For password hashing
python-dotenv
celery                   # For background tasks
redis                    # For Celery broker (we'll add redis to docker later if needed)
requests
tenacity                 # For retrying connections
Pillow>=10.0.0           # For Image Metadata
python-docx>=0.8.11      # For Word Doc Metadata
mutagen>=1.46.0          # For Audio/Video Metadata
chromadb>=0.4.0          # Vector Database for RAG
sentence-transformers>=2.2.0 # For Local Embeddings
llama-index>=0.10.0
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
passlib[bcrypt]          # Password Hashing
python-jose[cryptography] # JWT Tokens
python-multipart         # For OAuth2 Form Data
